+++ materialSync/src/add.ts
@@ -1,17 +1,16 @@
 import fs from 'fs';
-import path from 'path';
 import { exit } from 'process';
 import minimist from 'minimist';
 import {
     DOC_PATHS,
     loadMaterialConfig,
     relative2Doc,
+    REPO_ROOT,
     resolveMaterialConfig,
     saveMaterialConfig
-} from './material_helpers';
+} from './helpers';
 
-const repoRoot = path.resolve(__dirname, '..');
-process.chdir(repoRoot);
+process.chdir(REPO_ROOT);
 
 const configs = loadMaterialConfig();
 const argv = minimist(process.argv.slice(2));
