+++ materialSync/tsconfig.json
@@ -11,6 +11,6 @@
         "resolveJsonModule": true,
         "moduleResolution": "node"
     },
-    "include": ["*.ts"],
+    "include": ["src/*.ts"],
     "exclude": ["node_modules", "dist"]
 }
